requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
pydantic>=2.10.0
pydantic-settings>=2.7.0
langchain>=0.3.14
langchain-community>=0.3.14
langchain-openai>=0.3.0
langchain-huggingface>=0.1.2
sentence-transformers>=3.3.0
faiss-cpu>=1.9.0
pypdf>=5.1.0
python-docx>=1.1.2
websockets>=14.1
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
ruff>=0.8.0
